Convert MatchedClient tests to OCSClient tests
Note that all tests passed before this commit, when using the skeleton
MatchedClient.
